Create BaseTDOperator to extend in each TD operator
