+++ digdag-standards/src/main/java/io/digdag/standards/operator/td/TDOperator.java
@@ -7,7 +7,6 @@ import com.google.common.base.Throwables;
 import com.treasuredata.client.TDClient;
 import com.treasuredata.client.TDClientException;
 import com.treasuredata.client.TDClientHttpConflictException;
-import com.treasuredata.client.TDClientHttpException;
 import com.treasuredata.client.TDClientHttpNotFoundException;
 import com.treasuredata.client.model.TDJob;
 import com.treasuredata.client.model.TDJobRequest;
@@ -19,9 +18,7 @@ import io.digdag.spi.TaskExecutionException;
 import io.digdag.standards.operator.DurationInterval;
 import io.digdag.standards.operator.state.TaskState;
 import io.digdag.util.DurationParam;
-import io.digdag.util.RetryExecutor;
 import io.digdag.util.RetryExecutor.RetryGiveupException;
-import org.eclipse.jetty.http.HttpStatus;
 import org.immutables.value.Value;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,12 +27,10 @@ import java.io.Closeable;
 import java.time.Duration;
 import java.util.Map;
 import java.util.UUID;
-import java.util.concurrent.Callable;
 
 import static com.treasuredata.client.model.TDJob.Status.SUCCESS;
-import static io.digdag.util.RetryExecutor.retryExecutor;
 
-public class TDOperator
+public class TDOperator extends BaseTDOperator
         implements Closeable
 {
     public interface SystemDefaultConfig
@@ -46,11 +41,6 @@ public class TDOperator
     private static final Logger logger = LoggerFactory.getLogger(TDOperator.class);
 
     // TODO: adjust these retry intervals and limits when all td operators have persistent retry mechanisms implemented
-
-    private static final int INITIAL_RETRY_WAIT = 500;
-    private static final int MAX_RETRY_WAIT = 2000;
-    private static final int MAX_RETRY_LIMIT = 3;
-    public static final int AUTH_MAX_RETRY_LIMIT = 1;
 
     public static TDOperator fromConfig(BaseTDClientFactory clientFactory, SystemDefaultConfig systemDefaultConfig, Map<String, String> env, Config config, SecretProvider secrets)
     {
@@ -71,18 +61,6 @@ public class TDOperator
 
         return new TDOperator(client, database, secrets);
     }
-
-    public void updateApikey(SecretProvider secrets)
-    {
-        String apikey = TDClientFactory.getApikey(secrets);
-        client = client.withApiKey(apikey);
-    }
-
-    static final RetryExecutor defaultRetryExecutor = retryExecutor()
-            .withInitialRetryWait(INITIAL_RETRY_WAIT)
-            .withMaxRetryWait(MAX_RETRY_WAIT)
-            .withRetryLimit(MAX_RETRY_LIMIT)
-            .retryIf((exception) -> !isDeterministicClientException(exception));
 
     public static String escapeHiveIdent(String ident)
     {
@@ -116,9 +94,7 @@ public class TDOperator
         }
     }
 
-    private TDClient client;
     private final String database;
-    private final SecretProvider secrets;
 
     TDOperator(TDClient client, String database, SecretProvider secrets)
     {
@@ -135,34 +111,6 @@ public class TDOperator
     public String getDatabase()
     {
         return database;
-    }
-
-    public RetryExecutor authenticatinRetryExecutor() {
-        return retryExecutor()
-            .withInitialRetryWait(INITIAL_RETRY_WAIT)
-            .withMaxRetryWait(MAX_RETRY_WAIT)
-            .withRetryLimit(AUTH_MAX_RETRY_LIMIT)
-            .onRetry((exception, retryCount, retryLimit, retryWait) -> {
-                logger.debug("apikey will be tried to update by retrying");
-                updateApikey(secrets);
-            })
-            .retryIf((exception) -> isAuthenticationErrorException(exception));
-    }
-
-    private <T> T callWithRetry(Callable<T> op)
-    {
-        try {
-            return defaultRetryExecutor.run(() -> {
-                try {
-                    return authenticatinRetryExecutor().run(() -> op.call());
-                } catch (RetryGiveupException ex) {
-                    throw Throwables.propagate(ex.getCause());
-                }
-            });
-        }
-        catch (RetryGiveupException ex) {
-            throw Throwables.propagate(ex.getCause());
-        }
     }
 
     private void runWithRetry(Runnable op, Class ignoreException)
@@ -401,57 +349,6 @@ public class TDOperator
     private static int exponentialBackoffInterval(DurationInterval pollInterval, int iteration)
     {
         return (int) Math.min(pollInterval.min().getSeconds() * Math.pow(2, iteration), pollInterval.max().getSeconds());
-    }
-
-    static boolean isDeterministicClientException(Exception ex)
-    {
-        if (ex instanceof TDClientHttpException) {
-            int statusCode = ((TDClientHttpException) ex).getStatusCode();
-            switch (statusCode) {
-                case HttpStatus.TOO_MANY_REQUESTS_429:
-                case HttpStatus.REQUEST_TIMEOUT_408:
-                    return false;
-                default:
-                    // return true if 4xx
-                    return statusCode >= 400 && statusCode < 500;
-            }
-        }
-        return isFailedBeforeSendClientException(ex);
-    }
-
-    static boolean isAuthenticationErrorException(Exception ex)
-    {
-        if (ex instanceof TDClientHttpException) {
-            int statusCode = ((TDClientHttpException) ex).getStatusCode();
-            switch (statusCode) {
-                case HttpStatus.UNAUTHORIZED_401:
-                    // This is not for authentication basically, but it may be 403 for auth token error. https://tools.ietf.org/html/rfc6750
-                case HttpStatus.FORBIDDEN_403:
-                    return true;
-                default:
-                    return false;
-            }
-        }
-        else {
-            return false;
-        }
-    }
-
-    static boolean isFailedBeforeSendClientException(Exception ex)
-    {
-        if (ex instanceof TDClientException) {
-            // failed before sending HTTP request or receiving HTTP response
-            TDClientException.ErrorType errorType = ((TDClientException) ex).getErrorType();
-            switch (errorType) {
-                case INVALID_CONFIGURATION:  // failed to read td.conf, failed to pares integer in properties set to TDClientBuilder, etc.
-                case INVALID_INPUT:          // early table name validation fails, failed to format request body in json, etc.
-                    return true;
-                default:
-                    // other cases such as PROXY_AUTHENTICATION_FAILURE, SSL_ERROR, REQUEST_TIMEOUT, INTERRUPTED, etc.
-                    break;  // pass-through
-            }
-        }
-        return false;
     }
 
     @Override
